refactor: extract draggable to separate file
