+++ app/src/main/java/es/voghdev/hellocompose/draggableScreen.kt
@@ -1,135 +1,24 @@
 package es.voghdev.hellocompose
 
 import androidx.compose.foundation.*
-import androidx.compose.foundation.gestures.detectDragGesturesAfterLongPress
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Card
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.layout.boundsInWindow
-import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.rememberImagePainter
 
 data class Item(val title: String)
-
-internal class DragTargetInfo {
-    var isDragging: Boolean by mutableStateOf(false)
-    var dragPosition by mutableStateOf(Offset.Zero)
-    var dragOffset by mutableStateOf(Offset.Zero)
-    var draggableComposable by mutableStateOf<(@Composable () -> Unit)?>(null)
-    var dataToDrop by mutableStateOf<Any?>(null)
-}
-
-internal val LocalDragTargetInfo = compositionLocalOf { DragTargetInfo() }
-
-@Composable
-fun <T> DragTarget(
-    modifier: Modifier,
-    dataToDrop: T,
-    content: @Composable (() -> Unit)
-) {
-    var currentPosition by remember { mutableStateOf(Offset.Zero) }
-    val currentState = LocalDragTargetInfo.current
-
-    Box(modifier = modifier
-        .onGloballyPositioned {
-            currentPosition = it.localToWindow(Offset.Zero)
-        }
-        .pointerInput(Unit) {
-            detectDragGesturesAfterLongPress(onDragStart = {
-                currentState.dataToDrop = dataToDrop
-                currentState.isDragging = true
-                currentState.dragPosition = currentPosition + it
-                currentState.draggableComposable = content
-            }, onDrag = { change, dragAmount ->
-                change.consume()
-                currentState.dragOffset += Offset(dragAmount.x, dragAmount.y)
-            }, onDragEnd = {
-                currentState.isDragging = false
-                currentState.dragOffset = Offset.Zero
-            }, onDragCancel = {
-                currentState.dragOffset = Offset.Zero
-                currentState.isDragging = false
-            })
-        }) {
-        content()
-    }
-}
-
-@Composable
-fun <T> DropTarget(
-    modifier: Modifier,
-    content: @Composable() (BoxScope.(isInBound: Boolean, data: T?) -> Unit)
-) {
-
-    val dragInfo = LocalDragTargetInfo.current
-    val dragPosition = dragInfo.dragPosition
-    val dragOffset = dragInfo.dragOffset
-    var isCurrentDropTarget by remember {
-        mutableStateOf(false)
-    }
-
-    Box(modifier = modifier.onGloballyPositioned {
-        it.boundsInWindow().let { rect ->
-            isCurrentDropTarget = rect.contains(dragPosition + dragOffset)
-        }
-    }) {
-        val data =
-            if (isCurrentDropTarget && !dragInfo.isDragging) dragInfo.dataToDrop as T? else null
-        content(isCurrentDropTarget, data)
-    }
-}
-
-@Composable
-fun LongPressDraggable(
-    modifier: Modifier = Modifier,
-    content: @Composable BoxScope.() -> Unit
-) {
-    val state = remember { DragTargetInfo() }
-    CompositionLocalProvider(
-        LocalDragTargetInfo provides state
-    ) {
-        Box(modifier = modifier.fillMaxSize())
-        {
-            content()
-            if (state.isDragging) {
-                var targetSize by remember {
-                    mutableStateOf(IntSize.Zero)
-                }
-                Box(modifier = Modifier
-                    .graphicsLayer {
-                        val offset = (state.dragPosition + state.dragOffset)
-                        scaleX = 1.3f
-                        scaleY = 1.3f
-                        alpha = if (targetSize == IntSize.Zero) 0f else .9f
-                        translationX = offset.x.minus(targetSize.width / 2)
-                        translationY = offset.y.minus(targetSize.height / 2)
-                    }
-                    .onGloballyPositioned {
-                        targetSize = it.size
-                    }
-                ) {
-                    state.draggableComposable?.invoke()
-                }
-            }
-        }
-    }
-}
 
 @Composable
 fun DraggableScreen() {
@@ -150,56 +39,55 @@ fun DraggableScreen() {
 private fun DraggableItem(
     modifier: Modifier = Modifier,
     item: Item,
-) {
-    Box(modifier) {
-        HighlightedBackground()
-        Column(Modifier.clickable(onClick = { })) {
-            SmallSpacer()
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(
-                        bottom = 8.dp,
-                        start = 16.dp,
-                        end = 16.dp
-                    ),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                SampleImage()
-                SmallSpacer()
-                RowTitle(modifier = Modifier.weight(1f), text = item.title)
-                SmallSpacer()
-            }
-            Separator()
-        }
-        DragTarget(modifier = Modifier, dataToDrop = item) {
-            Image(
-                painter = painterResource(id = R.drawable.ic_launcher_foreground),
-                contentDescription = null,
-                contentScale = ContentScale.Crop,
-                modifier = Modifier
-                    .size(56.dp)
-                    .clip(RoundedCornerShape(16.dp))
-            )
-        }
-
-        DropTarget<Item>(
+) = Box(modifier) {
+    HighlightedBackground()
+    Column(Modifier.clickable(onClick = { })) {
+        SmallSpacer()
+        Row(
             modifier = Modifier
+                .fillMaxWidth()
+                .padding(
+                    bottom = 8.dp,
+                    start = 16.dp,
+                    end = 16.dp
+                ),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            SampleImage()
+            SmallSpacer()
+            RowTitle(modifier = Modifier.weight(1f), text = item.title)
+            SmallSpacer()
+        }
+        Separator()
+    }
+    DragTarget(modifier = Modifier, dataToDrop = item) {
+        Image(
+            painter = painterResource(id = R.drawable.ic_launcher_foreground),
+            contentDescription = null,
+            contentScale = ContentScale.Crop,
+            modifier = Modifier
+                .size(56.dp)
+                .clip(RoundedCornerShape(16.dp))
+        )
+    }
+
+    DropTarget<Item>(
+        modifier = Modifier
+            .size(24.dp)
+            .align(Alignment.CenterEnd)
+    ) { isInBound, item ->
+        val color = if (isInBound) {
+            Color.Red
+        } else {
+            Color.Gray
+        }
+        Box(
+            Modifier
                 .size(24.dp)
                 .align(Alignment.CenterEnd)
-        ) { isInBound, item ->
-            val color = if (isInBound) {
-                Color.Blue
-            } else {
-                Color.DarkGray
-            }
-            Box(
-                Modifier
-                    .size(24.dp)
-                    .align(Alignment.CenterEnd)
-                    .background(color)
-            )
-        }
+                .clip(RoundedCornerShape(16.dp))
+                .background(color)
+        )
     }
 }
 
@@ -220,39 +108,33 @@ private fun RowTitle(
     text: String,
     modifier: Modifier = Modifier,
     color: Color = Color.Black
-) {
-    Text(
-        text = text,
-        style = MaterialTheme.typography.body2,
-        color = color,
-        modifier = modifier,
-        lineHeight = 22.sp,
-        overflow = TextOverflow.Ellipsis
-    )
-}
+) = Text(
+    text = text,
+    style = MaterialTheme.typography.body2,
+    color = color,
+    modifier = modifier,
+    lineHeight = 22.sp,
+    overflow = TextOverflow.Ellipsis
+)
 
 @Composable
-private fun Separator() {
-    Box(
-        Modifier
-            .fillMaxWidth()
-            .padding(
-                start = 8.dp,
-                end = 8.dp
-            )
-            .height(1.dp)
-            .background(MaterialTheme.colors.onPrimary)
-    )
-}
+private fun Separator() = Box(
+    Modifier
+        .fillMaxWidth()
+        .padding(
+            start = 8.dp,
+            end = 8.dp
+        )
+        .height(1.dp)
+        .background(MaterialTheme.colors.onPrimary)
+)
 
 @Composable
-private fun HighlightedBackground() {
-    Card(
-        backgroundColor = Color.LightGray,
-        shape = RoundedCornerShape(16.dp),
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(64.dp)
-            .padding(start = 12.dp, end = 12.dp)
-    ) {}
-}
+private fun HighlightedBackground() = Card(
+    backgroundColor = Color.LightGray,
+    shape = RoundedCornerShape(16.dp),
+    modifier = Modifier
+        .fillMaxWidth()
+        .height(64.dp)
+        .padding(start = 12.dp, end = 12.dp)
+) {}
